docs/requirements.txt
myst-parser==4.0.0
sphinx_rtd_theme==2.0.0
sphinxcontrib-phpdomain==0.12.0